+++ cmd/fuse/main.go
@@ -10,7 +10,7 @@ import (
 	"github.com/spf13/cobra"
 )
 
-var version = "0.11.0"
+var version = "0.13.0"
 
 func main() {
 	root := &cobra.Command{